Allow first choice to contain leading whitespace
My use case presents choose with a list of options, where each option is a row
containing a few different columns. These columns can be right aligned, and are
separated by one or more spaces within a row.

This way, each row is just a string, but choose can still present to the user a
set of options where each option contains a bunch of neatly arranged columns.
This works so long as the string representation of each row is properly
formatted -- with the right width and alignment for each column and so forth.

I found that when the first column is right-aligned, it can happen that the
first row starts with a space. But choose will get rid of all such leading
spaces, thereby messing with my neatly arranged columns.

This commit is a one-line fix for this.