Add commandline switch: auto-select single choice
After filtering the available choices with the query string, sometimes it may
turn out that only 1 choice remains. In this case, one may want choose to just
select this one choice without waiting for the user to hit Enter.

This commit adds a command-line switch (-1) to the code that implements this
behavior.