Do not select initial query text
When we supply an initial query, due to the way macOS's text fields work, this
initial query is selected. So, any subsequent text we type ends up replacing the
initial query rather than appending to it.

This commit makes sure that the above does not happen. When an initial query is
present, any subsequent keypresses now have the effect of appending to said
query rather than replacing it.