Add option to match words instead of characters
Previously, choose could only match each element against the query character by
character. That is, if all characters in the query appeared in the element (in
order), the element was considered a match. Even if the characters didn't appear
in sequence.

For example, the row "banana" would match the query "bn" because the string
"banana" contains the string "bn" as a sub-sequence -- even though "b" and "n"
don't appear as consecutive characters anywhere in banana.

This commit adds an option (-W) to only consider rows as matching if the words
in the query appear in them as *continuous* sub-strings (NOT subsequences) -- in
order.

For example, with this option enabled, "banana" would match "an" but not "bn".
And "the quick brown fox" would match "qui fox" but not "qi fx" or "fox qui".