Support case-sensitive and smart-case matching
At the moment, choose does not allow us to match against queries
case-sensitively. It also does not allow smart-case (ie, case-sensitive when the
query contains uppercase characters, but case-insensitive if the query is
all-lowercase).

With this commit, we support these matching options. Specifically, I added a
command-line switch (-C) that takes 1 argument:

    - '-C i' = case-insensitive (the default),
    - '-C I' = case-sensitive, and
    - '-C s' = smart-case.

Also, I refactored some of the code for performing searching/matching. I think I
made the code a little easier to read -- especially around the parts that handle
matching from the beginning vs matching from the end. Also, I think I may have
improved performance a little: the way the code works, the same query is run
against all our choices. So, if some operation needs to be done on the query
(like splitting it into words/tokens), it's enough if we do this operation once,
and re-use the result for all our choices. That's what I'm doing in this commit.